% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spe_metadata3D.R
\name{add_spe_metadata3D}
\alias{add_spe_metadata3D}
\title{Add spaSim3D metadata to a SpatialExperiment object.}
\usage{
add_spe_metadata3D(spe, spe_metadata, plot_image = TRUE)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of
simulate_spe_metadata3D. This is because the metadata of the
SpatialExperiment object needs to already contain spaSim3D specific data
relating to the background of the SpatialExperiment object, and any
clusters.}

\item{spe_metadata}{A list containing the new metadata to add to the
SpatialExperiment object. This is generated from the output of the
spaSim3D function 'spe_metadata_cluster_template'. Avoid using the
'spe_metadata_background_template' as 'add_spe_metadata3D' is not
intended to be used to replace/update the old background metadata.}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
the added metadata. Defaults to TRUE.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the added metadata.
}
\description{
This function adds metadata to a SpatialExperiment object
generated by spaSim3D. This updates the metadata in the
SpatialExperiment object, giving it a new set of simulated clusters.
}
\examples{
# Get background metadata
bg_metadata <- spe_metadata_background_template("random")

# Get cluster metadata (using background metadata as background)
cluster_metadata <- spe_metadata_cluster_template("regular", "sphere", bg_metadata)
cluster_metadata <- spe_metadata_cluster_template("ring", "ellipsoid", cluster_metadata)

# Get spe from cluster metadata
spe_clusters <- simulate_spe_metadata3D(cluster_metadata)

# Get new metadata to add to spe
new_metadata <- spe_metadata_cluster_template("regular", "network")
new_metadata <- spe_metadata_cluster_template("double ring", "cylinder", new_metadata)

# Add new metadata to spe
spe_clusters1 <- add_spe_metadata3D(spe_clusters, new_metadata)

}
