% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells3D.R
\name{plot_cells3D}
\alias{plot_cells3D}
\title{Function to plot cells in 3D spatial data.}
\usage{
plot_cells3D(
  spe,
  plot_cell_types = NULL,
  plot_colours = NULL,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the viridis color pallete will be used.
Defaults to NULL.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}
}
\value{
A Plotly object plotting the cells of the 3D SpatialExperiment
Object.
}
\description{
This function plots the cells of 3D spatial data in a 3D
SpatialExperiment Object.
}
\examples{
# Get background metadata
bg_metadata <- spe_metadata_background_template("random")

# Get cluster metadata (using background metadata as background)
cluster_metadata <- spe_metadata_cluster_template("regular", "sphere", bg_metadata)
cluster_metadata <- spe_metadata_cluster_template("ring", "ellipsoid", cluster_metadata)

# Get spe from cluster metadata
spe_clusters <- simulate_spe_metadata3D(cluster_metadata, plot_image = FALSE)

# Plot
fig <- plot_cells3D(
    spe = spe_clusters,
    plot_cell_types = NULL,
    plot_colours = NULL,
    feature_colname = "Cell.Type"
)

methods::show(fig)

}
