% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_background_interactive.R
\name{simulate_background_interactive}
\alias{simulate_background_interactive}
\title{Function to simulate background of 3D tissue interactively.}
\usage{
simulate_background_interactive()
}
\value{
A 3D SpatialExperiment object with the background cells.
}
\description{
This is an interactive function that allows users to simulate
background cells of a 3D tissue. The function will ask you for your
inputs, for which the user can enter them in the console.
}
\examples{
# Simulate background
# background_spe <- simulate_background_interactive()

}
