% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cluster_interactive.R
\name{simulate_cluster_interactive}
\alias{simulate_cluster_interactive}
\title{Function to simulate clusters in a 3D tissue interactively.}
\usage{
simulate_cluster_interactive(simulated_spe = NULL)
}
\arguments{
\item{simulated_spe}{A SpatialExperiment object containing 3D spatial information
for the cells. The intended input should be generated using the output of
simulate_background_interactive or simulate_cluster_interactive, however,
any of the simulate_* functions should work. This is because the metadata
of the SpatialExperiment object needs to already contain spaSim3D
specific data relating to the background of the SpatialExperiment object,
and any clusters. Defaults to NULL, although, a valid spaSim3D
SpatialExperiment object is required.}
}
\value{
A 3D SpatialExperiment object with the new cell clusters.
}
\description{
This is an interactive function that allows users to simulate
cell clusters in a 3D tissue. The function will ask you for your
inputs, for which the user can enter them in the console.
}
\examples{
# Simulate background
# background_spe <- simulate_background_interactive()

# Simulate clusters
# clusters_spe <- simulate_cluster_interactive(simulated_spe = background_spe)

}
