% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_clusters3D.R
\name{simulate_clusters3D}
\alias{simulate_clusters3D}
\title{Simulate clusters in spaSim3D.}
\usage{
simulate_clusters3D(
  spe,
  cluster_properties_list,
  plot_image = TRUE,
  plot_cell_types = NULL,
  plot_colours = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{cluster_properties_list}{A list of lists containing the cluster
properties of each cluster desired. See the example to know what
properties are required for each cluster shape. If you want more detail
on each property, I recommend viewing the other simulate_* functions for
each particular shape.}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
the added metadata. Defaults to TRUE.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the viridis color pallete will be used.
Defaults to NULL.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new clusters and corresponding metadata.
}
\description{
This function simulates clusters specified by the user onto an
existing SpatialExperiment object. This includes customisable spheres,
ellipsoids, cylinders and network clusters.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)
                                           
# Simulate clusters
clusters <- simulate_clusters3D(bg_r,
                                cluster_properties = list(
                                  # Sphere example
                                  C1 = list(
                                    shape = "sphere",
                                    cluster_cell_types = c("Tumour", "Immune", "Others"),
                                    cluster_cell_proportions = c(0.55, 0.4, 0.05),
                                    radius = 25,
                                    centre_loc = c(40, 40, 40)
                                  ),
                                  # Cylinder example
                                  C2 = list(
                                    shape = "cylinder",
                                    cluster_cell_types = c("Endothelial", "Others"),
                                    cluster_cell_proportions = c(0.95, 0.05),
                                    radius = 10,
                                    start_loc = c(0, 0, 0),
                                    end_loc   = c(20, 20 , 100)
                                  ),
                                  # Ellipsoid example
                                  C3 = list(
                                    shape = "ellipsoid",
                                    cluster_cell_types = c("Tumour", "Immune", "Others"),
                                    cluster_cell_proportions = c(0.65, 0.3, 0.05),
                                    radii = c(15, 20, 25),
                                    centre_loc = c(70, 70, 70),
                                    axes_rotation = c(0, 0, 0)
                                  )
                                  # Network example
                                  C4 = list(
                                    shape = "network",
                                    cluster_cell_types = c("Immune"),
                                    cluster_cell_proportions = c(1),
                                    n_edges = 15,
                                    width = 8,
                                    centre_loc = c(75, 75, 25), # Rough centre of network cluster
                                    radius = 40 # Rough radius spanned by the network cluster
                                  )
                                ),
                                plot_image = TRUE,
                                plot_cell_types = c("Others", "Immune", "Endothelial", "Tumour"),
                                plot_colours = c("lightgray", "skyblue", "#FF7F7F", "orange"))
                                           
}
