% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cylinder_cluster.R
\name{simulate_cylinder_cluster}
\alias{simulate_cylinder_cluster}
\title{Simulate a cylinder cluster in spaSim3D.}
\usage{
simulate_cylinder_cluster(spe, cluster_properties)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{cluster_properties}{A list containing the properties of the cylinder
cluster desired. The list should contain the following elements:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"shape": Must be equal to the character "cylinder".

"cluster_cell_types": A character vector representing the cell types that
    make up the cluster. E.g. c("Tumour", "Immune").

"cluster_cell_proportions": A numeric vector representing the proportion
    of each cell type in the cluster. Its elements must each be
    greater than 0, sum to 1 and the vector must be the same length as
    "cluster_cell_types". E.g. c(0.6, 0.4) corresponds to a cluster made
    up of 60\% Tumour and 40\% Immune.

"radius": A positive number representing the radius of the cylinder.

"start_loc": A numerical vector of length 3 representing the starting
    x,y,z coordinate of the cylinder. E.g. (0, 0, 0).

"end_loc": A numerical vector of length 3 representing the ending x,y,z
    coordinate of the cylinder. E.g. (20, 20, 20).
}\if{html}{\out{</div>}}}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new cylinder cluster and the corresponding metadata.
}
\description{
This function simulates a cylinder cluster onto an existing
SpatialExperiment object. The parameters of the cylinder are fully
customisable by the user.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)

# Simulate cluster
cylinder_cluster <- simulate_cylinder_cluster(bg_r,
                                              cluster_properties = list(
                                                shape = "cylinder",
                                                cluster_cell_types = c("Endothelial", "Others"),
                                                cluster_cell_proportions = c(0.95, 0.05),
                                                radius = 10,
                                                start_loc = c(0, 0, 0),
                                                end_loc = c(20, 20, 100)
                                              ))
# Plot
plot_cells3D(cylinder_cluster)

}
