% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ellipsoid_ring.R
\name{simulate_ellipsoid_ring}
\alias{simulate_ellipsoid_ring}
\title{Simulate an ellipsoid cluster with a ring in spaSim3D.}
\usage{
simulate_ellipsoid_ring(spe, ring_properties)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{ring_properties}{A list containing the properties of the ellipsoid
cluster and ring desired. The list should contain the following elements:
"shape": Must be equal to the character "ellipsoid".
"cluster_cell_types": A character vector representing the cell types that
make up the cluster. E.g. c("Tumour", "Immune").
"cluster_cell_proportions": A numeric vector representing the proportion
of each cell type in the cluster. Its elements must each be
greater than 0, sum to 1 and the vector must be the same length as
"cluster_cell_types". E.g. c(0.6, 0.4) corresponds to a cluster made
up of 60\% Tumour and 40\% Immune.
"radii": A numerical vector of length 3 containing only positive numbers,
representing the radii of the ellipsoid in the x,y,z direction
respectively. E.g. c(15, 20, 25).
"centre_loc": A numerical vector of length 3 representing the centre
x,y,z coordinate of the ellipsoid E.g. (0, 0, 0).
"axes_rotation": A numerical vector of length 3 representing axes
rotation of the ellipsoid in the y-z plane, x-z plane and x-y plane
respectively. Values should be in degrees unit. E.g. (30, 45, 0).
"ring_cell_types": A character vector representing the cell types
that make up the ring. E.g. c("Immune1", "Immune2", "Immune3").
"ring_cell_proportions": A numeric vector representing the
proportion of each cell type in the ring. Its elements must each be
greater than 0, sum to 1 and the vector must be the same length as
"ring_cell_types". E.g. c(0.3, 0.4, 0.3) corresponds to an
ring made up of 30\% Immune1, 40\% Immune2 and 30\% Immune3.
"ring_width": A positive number representing the width of the ring.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new ellipsoid with ring cluster and the corresponding metadata.
}
\description{
This function simulates an ellipsoid cluster with a ring onto an
existing SpatialExperiment object. The parameters of the ellipsoid and
its ring are fully customisable by the user.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)

# Simulate cluster
ellipsoid_ring_cluster <- simulate_ellipsoid_ring(bg_r,
                                                  ring_properties = list(
                                                    shape = "ellipsoid",
                                                    cluster_cell_types = c("Tumour", "Others"),
                                                    cluster_cell_proportions = c(0.95, 0.05),
                                                    radii = c(20, 30, 40),
                                                    centre_loc = c(70, 70, 70),
                                                    axes_rotation = c(0, 0, 0),
                                                    ring_cell_types = c("Immune", "Others"),
                                                    ring_cell_proportions = c(0.85, 0.15),
                                                    ring_width = 5
                                                  ))
# Plot
plots_cells3D(ellipsoid_ring_cluster)

}
