% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mixing3D.R
\name{simulate_mixing3D}
\alias{simulate_mixing3D}
\title{Simulate mixing in spaSim3D.}
\usage{
simulate_mixing3D(
  spe,
  cell_types,
  cell_proportions,
  plot_image = TRUE,
  plot_cell_types = NULL,
  plot_colours = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{cell_types}{A character vector representing the cell types that will
make up the background of the SpatialExperiment object. E.g.
c("Tumour", "Immune").}

\item{cell_proportions}{A numeric vector representing the proportion
of each cell type in the background of the SpatialExperiment object.
Its elements must each be greater than 0, sum to 1 and the vector
must be the same length as "cell_types". E.g. c(0.6, 0.4) corresponds
to a background made up of 60\% Tumour and 40\% Immune.}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
the added metadata. Defaults to TRUE.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the viridis color pallete will be used.
Defaults to NULL.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new mixing of the background and corresponding metadata.
}
\description{
This function simulates mixing of an existing SpatialExperiment
object. This means completely updating the cell types that make up the
background of the SpatialExperiment object.
}
\examples{
# Simulate background
bg_o <- simulate_ordered_background_cells3D(n_cells = 10000,
                                            length = 100,
                                            width = 100,
                                            height = 100,
                                            jitter_proportion = 0,
                                            background_cell_type = "Others",
                                            plot_image = FALSE)

# Simulate mixing
bg_mix <- simulate_mixing3D(bg_o,
                            cell_types = c("Others", "Immune", "Tumour"),
                            cell_proportions = c(0.5, 0.25, 0.25),
                            plot_image = TRUE,
                            plot_cell_types = c("Others", "Immune", "Tumour"),
                            plot_colours = c("lightgray", "skyblue", "orange"))

# Re-plotting so it shows in GitHub pages
plot_cells3D(bg_mix,
             plot_cell_types = c("Others", "Immune", "Tumour"),
             plot_colours = c("lightgray", "skyblue", "orange"))

}
