% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_network_dr.R
\name{simulate_network_dr}
\alias{simulate_network_dr}
\title{Simulate a network cluster with a double ring in spaSim3D.}
\usage{
simulate_network_dr(spe, dr_properties)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{dr_properties}{A list containing the properties of the network cluster
and double ring desired. The list should contain the following elements:
"shape": Must be equal to the character "network".
"cluster_cell_types": A character vector representing the cell types that
make up the cluster. E.g. c("Tumour", "Immune").
"cluster_cell_proportions": A numeric vector representing the proportion
of each cell type in the cluster. Its elements must each be
greater than 0, sum to 1 and the vector must be the same length as
"cluster_cell_types". E.g. c(0.6, 0.4) corresponds to a cluster made
up of 60\% Tumour and 40\% Immune.
"n_edges": A positive number capped at 99 representing the number of
edges.
"width": A positive number representing the width of each network edge.
"centre_loc": A numerical vector of length 3 representing the approximate
centre x,y,z coordinate of the network E.g. (0, 0, 0).
"radius": A positive number representing the rough radius spanned by the
network.
"inner_ring_cell_types": A character vector representing the cell types
that make up the inner ring. E.g. c("Immune1", "Immune2", "Immune3").
"inner_ring_cell_proportions": A numeric vector representing the
proportion of each cell type in the inner ring. Its elements must
each be greater than 0, sum to 1 and the vector must be the same
length as "inner_ring_cell_types". E.g. c(0.3, 0.4, 0.3) corresponds
to an inner ring made up of 30\% Immune1, 40\% Immune2 and 30\% Immune3.
"inner_ring_width": A positive number representing the width of the inner
ring.
"outer_ring_cell_types": A character vector representing the cell types
that make up the outer ring. E.g. c("T cell", "B cell").
"outer_ring_cell_proportions": A numeric vector representing the
proportion of each cell type in the outer ring. Its elements must
each be greater than 0, sum to 1 and the vector must be the same
length as "outer_ring_cell_types". E.g. c(0.5, 0.5) corresponds to an
outer ring made up of 50\% T cell and 50\% B cell.
"outer_ring_width": A positive number representing the width of the outer
ring.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new network with double ring cluster and the corresponding
metadata.
}
\description{
This function simulates a network cluster with a double ring
onto an existing SpatialExperiment object. The parameters of the network
and its double ring are fully customisable by the user.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)

# Simulate cluster
network_dr_cluster <- simulate_network_dr(bg_r,
                                          dr_properties = list(
                                            shape = "network",
                                            cluster_cell_types = c("Tumour"),
                                            cluster_cell_proportions = c(1),
                                            n_edges = 15,
                                            width = 5,
                                            centre_loc = c(50, 50, 50),
                                            radius = 50,
                                            inner_ring_cell_types = c("Endothelial"),
                                            inner_ring_cell_proportions = c(1),
                                            inner_ring_width = 2,
                                            outer_ring_cell_types = c("Immune"),
                                            outer_ring_cell_proportions = c(1),
                                            outer_ring_width = 2
                                          ))
# Plot
plots_cells3D(network_dr_cluster)

}
