% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ordered_background_cells3D.R
\name{simulate_ordered_background_cells3D}
\alias{simulate_ordered_background_cells3D}
\title{Simulate ordered background cells in spaSim3D.}
\usage{
simulate_ordered_background_cells3D(
  n_cells,
  length,
  width,
  height,
  jitter_proportion = 0.25,
  background_cell_type = "Others",
  plot_image = TRUE
)
}
\arguments{
\item{n_cells}{A positive number representing the number of cells in the
background.}

\item{length}{A positive number representing the length of the 3D window of
the background. E.g. If you want a 100 by 150 by 200 unit window, set
length to 100.}

\item{width}{A positive number representing the width of the 3D window of
the background. E.g. If you want a 100 by 150 by 200 unit window, set
width to 150.}

\item{height}{A positive number representing the height of the 3D window of
the background. E.g. If you want a 100 by 150 by 200 unit window, set
height to 200.}

\item{jitter_proportion}{A number between 0 and 1 representing the amount of
'jitter' to apply to all cells. If jitter_proportion = 0.5 and the
original distance between each cell is 5 units, then each cell will move
UP TO 0.5 * 5 = 2.5 units away from its original position in the x, y and
z directions. The 'original distance between each cell' is calculated by
this function and will depend on the previous parameters, not something
to worry about :)}

\item{background_cell_type}{A character representing the cell type label of
all the cells in the background. Defaults to "Others".}

\item{plot_image}{A logical indicating whether to plot 3D spatial data.
Defaults to TRUE.}
}
\value{
A 3D SpatialExperiment object with the background cells and
corresponding metadata.
}
\description{
This function simulates background cells in an 'ordered'
fashion, so that cells are geometrically equidistant apart.
The parameters of the background are completely customisable by the user.
}
\examples{
# Simulate background
bg_o <- simulate_ordered_background_cells3D(n_cells = 10000,
                                            length = 100,
                                            width = 100,
                                            height = 100,
                                            jitter_proportion = 0,
                                            background_cell_type = "Others",
                                            plot_image = TRUE)

# Re-plotting so it shows in GitHub pages
plot_cells3D(bg_o,
             plot_cell_types = "Others",
             plot_colours = "lightgray")

}
