% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rings3D.R
\name{simulate_rings3D}
\alias{simulate_rings3D}
\title{Simulate clusters with rings in spaSim3D.}
\usage{
simulate_rings3D(
  spe,
  ring_properties_list,
  plot_image = TRUE,
  plot_cell_types = NULL,
  plot_colours = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{ring_properties_list}{A list of lists containing the cluster and ring
properties of each cluster desired. See the example to know what
properties are required for each cluster shape. If you want more detail
on each property, I recommend viewing the other simulate_* functions for
each particular shape.}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
the added metadata. Defaults to TRUE.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the viridis color pallete will be used.
Defaults to NULL.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new cluster with rings and corresponding metadata.
}
\description{
This function simulates clusters specified by the user onto an
existing SpatialExperiment object. This includes customisable spheres,
ellipsoids, cylinders and network clusters, all with rings.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)
                                           
# Simulate clusters with rings
clusters_with_rings <- simulate_rings3D(bg_r,
                                        ring_properties = list(
                                          R1 = list(
                                            shape = "sphere",
                                            cluster_cell_types = c("Tumour", "Others"),
                                            cluster_cell_proportions = c(0.95, 0.05),
                                            radius = 20,
                                            centre_loc = c(40, 40, 40),
                                            ring_cell_types = c("Immune", "Others"),
                                            ring_cell_proportions = c(0.85, 0.15),
                                            ring_width = 5
                                          ),
                                          R2 = list(
                                            shape = "cylinder",
                                            cluster_cell_types = c("Void"),
                                            cluster_cell_proportions = c(1),
                                            radius = 8,
                                            start_loc = c(0, 0, 0),
                                            end_loc   = c(20, 20 , 100),
                                            ring_cell_types = c("Endothelial", "Others"),
                                            ring_cell_proportions = c(0.85, 0.15),
                                            ring_width = 5
                                          ),
                                          R3 = list(
                                            shape = "ellipsoid",
                                            cluster_cell_types = c("Tumour", "Others"),
                                            cluster_cell_proportions = c(0.95, 0.05),
                                            radii = c(20, 30, 40),
                                            centre_loc = c(70, 70, 70),
                                            axes_rotation = c(0, 0, 0),
                                            ring_cell_types = c("Immune", "Others"),
                                            ring_cell_proportions = c(0.85, 0.15),
                                            ring_width = 5
                                          ),
                                          R4 = list(
                                            shape = "network",
                                            cluster_cell_types = c("Tumour"),
                                            cluster_cell_proportions = c(1),
                                            n_edges = 15,
                                            width = 5,
                                            centre_loc = c(75, 75, 25),
                                            radius = 50,
                                            ring_cell_types = c("Endothelial"),
                                            ring_cell_proportions = c(1),
                                            ring_width = 2
                                          )
                                        ),
                                        plot_image = TRUE,
                                        plot_cell_types = c("Others", "Tumour", "Immune", "Endothelial"),
                                        plot_colours = c("lightgray", "orange", "skyblue", "#FF7F7F"))
                                           
}
