% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spe_metadata3D.R
\name{simulate_spe_metadata3D}
\alias{simulate_spe_metadata3D}
\title{Simulate a SpatialExperiment object with spaSim3D metadata.}
\usage{
simulate_spe_metadata3D(
  spe_metadata,
  plot_image = TRUE,
  plot_cell_types = NULL,
  plot_colours = NULL
)
}
\arguments{
\item{spe_metadata}{A list containing the metadata to generate the
SpatialExperiment object. This is generated from the output of the
spaSim3D functions 'spe_metadata_background_template' and
'spe_metadata_cluster_template'.}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
the added metadata. Defaults to TRUE.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the viridis color pallete will be used.
Defaults to NULL.}
}
\value{
A SpatialExperiment object containing the cells of the simulated 3D
tissue and including the inputted metadata.
}
\description{
This function generates a SpatialExperiment object with
parameters specified with inputted spaSim3D metadata. By doing so, it
simulates a 3D tissue, beginning with the background cells of the tissue,
then adding cell clusters to the tissue.
}
\examples{
# Get background metadata
bg_metadata <- spe_metadata_background_template("random")

# Get cluster metadata (using background metadata as background)
cluster_metadata <- spe_metadata_cluster_template("regular", "sphere", bg_metadata)
cluster_metadata <- spe_metadata_cluster_template("ring", "ellipsoid", cluster_metadata)

# Get spe from cluster metadata
spe_clusters <- simulate_spe_metadata3D(cluster_metadata)

}
