% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sphere_dr.R
\name{simulate_sphere_dr}
\alias{simulate_sphere_dr}
\title{Simulate a sphere cluster with a double ring in spaSim3D.}
\usage{
simulate_sphere_dr(spe, dr_properties)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Should be generated using the output of one of the following
functions: simulate_random_background_cells3D,
simulate_ordered_background_cells3D,
simulate_mixing3D or any of the other simulate_* functions. This is
because the metadata of the SpatialExperiment object needs to already
contain spaSim3D specific data relating to the background of the
SpatialExperiment object, and any clusters.}

\item{dr_properties}{A list containing the properties of the sphere cluster
and double ring desired. The list should contain the following elements:
"shape": Must be equal to the character "sphere".
"cluster_cell_types": A character vector representing the cell types that
make up the cluster. E.g. c("Tumour", "Immune").
"cluster_cell_proportions": A numeric vector representing the proportion
of each cell type in the cluster. Its elements must each be
greater than 0, sum to 1 and the vector must be the same length as
"cluster_cell_types". E.g. c(0.6, 0.4) corresponds to a cluster made
up of 60\% Tumour and 40\% Immune.
"radius": A positive number representing the radius of the sphere.
"centre_loc": A numerical vector of length 3 representing the centre
x,y,z coordinate of the sphere E.g. (40, 50, 60).
"inner_ring_cell_types": A character vector representing the cell types
that make up the inner ring. E.g. c("Immune1", "Immune2", "Immune3").
"inner_ring_cell_proportions": A numeric vector representing the
proportion of each cell type in the inner ring. Its elements must
each be greater than 0, sum to 1 and the vector must be the same
length as "inner_ring_cell_types". E.g. c(0.3, 0.4, 0.3) corresponds
to an inner ring made up of 30\% Immune1, 40\% Immune2 and 30\% Immune3.
"inner_ring_width": A positive number representing the width of the inner
ring.
"outer_ring_cell_types": A character vector representing the cell types
that make up the outer ring. E.g. c("T cell", "B cell").
"outer_ring_cell_proportions": A numeric vector representing the
proportion of each cell type in the outer ring. Its elements must
each be greater than 0, sum to 1 and the vector must be the same
length as "outer_ring_cell_types". E.g. c(0.5, 0.5) corresponds to an
outer ring made up of 50\% T cell and 50\% B cell.
"outer_ring_width": A positive number representing the width of the outer
ring.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, updated
with the new sphere with double ring cluster and the corresponding
metadata.
}
\description{
This function simulates a sphere cluster with a double ring
onto an existing SpatialExperiment object. The parameters of the sphere
and its double ring are fully customisable by the user.
}
\examples{
# Simulate background
bg_r <- simulate_random_background_cells3D(n_cells = 10000,
                                           length = 100,
                                           width = 100,
                                           height = 100,
                                           minimum_distance_between_cells = 0.5,
                                           background_cell_type = "Others",
                                           plot_image = FALSE)
                                           
# Simulate cluster
sphere_dr_cluster <- simulate_cylinder_dr(bg_r,
                                          dr_properties = list(
                                            shape = "sphere",
                                            cluster_cell_types = c("Tumour", "Others"),
                                            cluster_cell_proportions = c(0.95, 0.05),
                                            radius = 20,
                                            centre_loc = c(40, 40, 40),
                                            inner_ring_cell_types = c("Immune1", "Others"),
                                            inner_ring_cell_proportions = c(0.85, 0.15),
                                            inner_ring_width = 5,
                                            outer_ring_cell_types = c("Immune2"),
                                            outer_ring_cell_proportions = c(1),
                                            outer_ring_width = 3
                                          ))
# Plot
plots_cells3D(sphere_dr_cluster)
                                            
}
