% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaSim3D_background_integrator.R
\name{spaSim3D_background_integrator}
\alias{spaSim3D_background_integrator}
\title{Function to simulate background of 3D tissue.}
\usage{
spaSim3D_background_integrator()
}
\value{
A 3D SpatialExperiment object with the background cells.
}
\description{
This is an interactive function that allows users to simulate
background cells of a 3D tissue. The function will ask you for your
inputs, for which the user can enter them in the console.
}
\examples{
# Simulate background
background_spe <- spaSim3D_background_integrator()

}
