% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaSim3D_cluster_integrator.R
\name{spaSim3D_cluster_integrator}
\alias{spaSim3D_cluster_integrator}
\title{Function to simulate clusters in a 3D tissue.}
\usage{
spaSim3D_cluster_integrator(simulated_spe = NULL)
}
\value{
A 3D SpatialExperiment object with the new cell clusters.
}
\description{
This is an interactive function that allows users to simulate
cell clusters in a 3D tissue. The function will ask you for your
inputs, for which the user can enter them in the console.
}
\examples{
# Simulate background
background_spe <- spaSim3D_background_integrator()

# Simulate clusters
clusters_spe <- spaSim3D_cluster_integrator(simulated_spe = background_spe)

}
