% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_metadata_background_template.R
\name{spe_metadata_background_template}
\alias{spe_metadata_background_template}
\title{Create spaSim3D background metadata template.}
\usage{
spe_metadata_background_template(background_type, original_spe_metadata = NULL)
}
\arguments{
\item{background_type}{Either "random" or "ordered". These are the two types
of backgrounds available in spaSim3D, each with their own set of
parameters.}

\item{original_spe_metadata}{A list containing the metadata for a spaSim3D
SpatialExperiment object. This is generated from the output of the
spaSim3D function 'spe_metadata_cluster_template'. Defaults to NULL. If
NULL, the background metadata template will be returned. If original
metadata is inputted, the background element of the original metadata
will be replaced with the background metadata template. If the original
metadata didn't have a background element, a background element will be
added, containing the background metadata template.}
}
\value{
A list containing the background metadata template or the original
metadata updated with the background metadata template.
}
\description{
This function creates a spaSim3D background metadata template
with the predefined parameters required. Users can change the parameters
after creating the template. Can be used as input for the following
functions: spe_metadata_cluster_template, simulate_spe_metadata3D.
}
\examples{
# Get background metadata
bg_metadata <- spe_metadata_background_template("random")

# Change background metadata
bg_metadata$background$n_cells <- 25000

# Get spe from background metadata
spe_background <- simulate_spe_metadata3D(bg_metadata)

}
