% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_metadata_cluster_template.R
\name{spe_metadata_cluster_template}
\alias{spe_metadata_cluster_template}
\title{Create spaSim3D cluster metadata template.}
\usage{
spe_metadata_cluster_template(
  cluster_type,
  shape,
  original_spe_metadata = NULL
)
}
\arguments{
\item{cluster_type}{Either "regular", "ring" or "double ring". These are the
three types of cluster types available in spaSim3D, each with their own
set of parameters.}

\item{shape}{Either "sphere", "ellipsoid", "cylinder" or "network". These are
the four types of cluster shapes available in spaSim3D, each with their
own set of parameters.}

\item{original_spe_metadata}{A list containing the metadata for a spaSim3D
SpatialExperiment object. This is generated from the output of the
spaSim3D function 'spe_metadata_background_template',
'spe_metadata_cluster_template'. Defaults to NULL. If original metadata
is inputted, the cluster metadata template will be added to it. If NULL,
the cluster metadata template will be returned, without a background
element. Hence, it is permissible to generate metadata that does not
include a background element. This can be inputted into the
add_spe_metadata3D function, however, cannot be inputed into
simulate_spe_metadata3D function as this function requires a background
element.}
}
\value{
A list containing the cluster metadata template or the original
metadata updated with the cluster metadata template.
}
\description{
This function creates a spaSim3D cluster metadata template with
the predefined parameters required. Users can change the parameters after
creating the template. Can be used as input for the following functions:
spe_metadata_background_template, spe_metadata_cluster_template,
simulate_spe_metadata3D, add_spe_metadata3D.
}
\examples{
# Get background metadata
bg_metadata <- spe_metadata_background_template("random")

# Change background metadata
bg_metadata$background$n_cells <- 25000

# Cluster metadata (using background metadata as background)
cluster_metadata <- spe_metadata_cluster_template("regular", "sphere", bg_metadata)
cluster_metadata <- spe_metadata_cluster_template("ring", "ellipsoid", cluster_metadata)
cluster_metadata <- spe_metadata_cluster_template("double ring", "cylinder", cluster_metadata)

# Change cluster metadata
cluster_metadata$cluster_1$radius <- 120

# Get spe from updated metadata
spe_clusters <- simulate_spe_metadata3D(cluster_metadata)
plot_cells3D(spe_clusters,
             plot_cell_types = c("Others", "Tumour", "Immune", "Immune1", "Endothelial"),
             plot_colours = c("lightgray", "orange", "skyblue", "lightgreen", "tomato"))

}
